Exactitude lists, move higher level code to Ex
